Allow templating all parameters of the FIO workload
Instead of having hardcoded randread and randwrite FIO profiles, allow
to customize the whole fio profile through the API. Also, FIO now
outputs in json and we parse the metrics with jq instead of the
fiostats.sh script (so jq needs to be installed).

The metrics that were already extracted are still in place and we added
the iops, bandwidth and latency measurements as well.

For the randread and read IO kinds, we manually create the file on disk
based on the fio_file_size parameter. This allows to only create it once
and run multiple tests without having to recreate it every time.

Signed-off-by: Julien Desfossez <jdesfossez@digitalocean.com>